+++ atlas-content-modeler.php
@@ -53,10 +53,13 @@ function atlas_content_modeler_loader(): void {
 		'blueprints/import.php',
 		'blueprints/fetch.php',
 		'blueprints/export.php',
+		'api/crud-functions.php',
+		'api/validation-functions.php',
+		'class-validation-exception.php',
 	);
 
 	foreach ( $plugin_files as $file ) {
-			include_once ATLAS_CONTENT_MODELER_INCLUDES_DIR . $file;
+		include_once ATLAS_CONTENT_MODELER_INCLUDES_DIR . $file;
 	}
 
 	if ( defined( 'WP_CLI' ) && WP_CLI ) {