+++ CHANGELOG.md
@@ -5,6 +5,15 @@
 - Focus now moves to the new input field when adding a new repeating number, date or single line text field row.
 - Pressing enter in a repeating number, date or single line text field now moves focus to the next field, or adds a new field if focus is in the last field.
 
+### Added
+- Function insert_entry_model() for inserting a model entry and fields.
+- Function update_model_entry() for updating an existing model entry and fields.
+- Function replace_relationship() to replace/add model entry relationships.
+- Function add_relationship() to append model entry relationships.
+- Function get_relationship() to retrieve the relationship object.
+- Function get_model() to retrieve the model schema.
+- Function get_field() to retrieve the model field schema.
+
 ## 0.16.0 - 2022-04-06
 ### Added
 - The date field now has the “make this field repeatable” option to let publishers add multiple dates within each date field.